user api improvement
