+++ packages/headless/src/database/mongodb/user/user.ts
@@ -22,13 +22,4 @@ export class UserDatabase implements IUserDatabase {
   async updateUser(userId: string, user: User): Promise<User | null> {
     return await UserModel.findOneAndUpdate({ id: userId }, { $set: user }, { new: true }).lean().select('-password -_id').exec();
   }
-
-  async updateUserWithNewAddress(userId: string, user: User, address: Address): Promise<User | null> {
-    return await UserModel.findOneAndUpdate({ id: userId }, { $push: { addresses: address }, $set: user, }, { new: true }).lean().select('-password -_id').exec();
-  }
-
-  async updateUserAndAddress(userId: string, user: User, address: Address): Promise<User | null> {
-    console.log(user.addresses)
-    return await UserModel.findOneAndUpdate({ id: userId, 'addresses.id': address.id }, { $set: user }, { new: true }).lean().select('-password -_id').exec();
-  }
 }
