+++ packages/headless/src/helper/mailService/index.ts
@@ -1,18 +1,18 @@
 import { Injectable } from "@nestjs/common";
 import { IMailService } from "./mail.service.interface";
 import * as nodemailer from 'nodemailer';
-import { mailConfig } from 'config/mail';
+import { nodemailerConfig } from 'config/mail';
 
 @Injectable()
 export class MailService implements IMailService {
     async sendMail(email: string, subject: string, mailBody: string): Promise<Boolean | null> {
         const mailOptions = {
-            from: mailConfig.user,
+            from: nodemailerConfig.user,
             to: email,
             subject,
             html: mailBody
         }
-        const transporter = nodemailer.createTransport(mailConfig.options);
+        const transporter = nodemailer.createTransport(nodemailerConfig.options);
         try {
             const res = await transporter.sendMail(mailOptions);
             if (!res) return false;
