+++ packages/headless/src/modules/user/repositories/index.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { Address, UpdatedUser, User } from 'src/entity/user';
+import { Address, User } from 'src/entity/user';
 import { IUserDatabase } from './user.database.interface';
 
 @Injectable()
@@ -20,14 +20,6 @@ export class UserRepository {
 
   async updateUser(userId: string, user: User): Promise<User | null> {
     return await this.db.updateUser(userId, user);
-  }
-
-  async updateUserWithNewAddress(userId: string, user: User, address: Address): Promise<User | null> {
-    return await this.db.updateUserWithNewAddress(userId, user, address);
-  }
-
-  async updateUserAndAddress(userId: string, user: User, address: Address): Promise<User | null> {
-    return await this.db.updateUserAndAddress(userId, user, address);
   }
 
 }
