+++ packages/headless/src/modules/user/repositories/user.database.interface.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { Address, UpdatedUser, User } from 'src/entity/user';
+import { Address, User } from 'src/entity/user';
 
 @Injectable()
 export abstract class IUserDatabase {
@@ -7,6 +7,4 @@ export abstract class IUserDatabase {
   abstract findUser: (query: Record<string, string>) => Promise<User | null>;
   abstract getUserPassword(query: Record<string, string>): Promise<User | null>;
   abstract updateUser: (userId: string, user: User) => Promise<User | null>;
-  abstract updateUserWithNewAddress: (userId: string, user: User, address: Address) => Promise<User | null>;
-  abstract updateUserAndAddress: (userId: string, user: User, address: Address) => Promise<User | null>;
 }
