+++ packages/headless/src/modules/user/validations/password.validator.ts
@@ -3,5 +3,4 @@ import { ChangePassword } from 'src/entity/user';
 export const ChangePasswordSchema = Joi.object<ChangePassword, true>({
     currentPassword: Joi.string().min(6).required().messages({ 'string.min': 'Password must be minimum 6 Characters', }),
     newPassword: Joi.string().min(6).required().messages({ 'string.min': 'Password must be minimum 6 Characters', }),
-    verifyPassword: Joi.string().min(6).required().messages({ 'string.min': 'Password must be minimum 6 Characters', }),
 });
